Minor changes in controllers to handle errors

+++ BookshopApi/Controllers/BookingController.cs
@@ -12,11 +12,13 @@ namespace BookshopApi.Controllers;
 [ApiController]
 public class BookingController : ControllerBase
 {
-    private IBookingService _bookingService;
+    private readonly IBookingService _bookingService;
+    private readonly ILogger<BookingController> _logger;
 
-    public BookingController(IBookingService bookingService)
+    public BookingController(IBookingService bookingService, ILogger<BookingController> logger)
     {
         _bookingService = bookingService;
+        _logger = logger;
     }
 
     [HttpGet]
@@ -43,59 +45,106 @@ public class BookingController : ControllerBase
              
             return Ok(bookings);
         }
-        catch (Exception)
+        catch (Exception ex)
         {
-            return StatusCode(StatusCodes.Status500InternalServerError, "An error occurred while getting the booking.");
+            _logger.LogError(ex, "An error occurred while getting the all bookings.");
+            return StatusCode(StatusCodes.Status500InternalServerError, ex.Message);
         }
     }
 
     [HttpGet]
     [ProducesResponseType(StatusCodes.Status200OK)]
     [ProducesResponseType(StatusCodes.Status500InternalServerError)]
+    [ProducesResponseType(StatusCodes.Status404NotFound)]
     [ProducesResponseType(StatusCodes.Status403Forbidden)]
+    [ProducesResponseType(StatusCodes.Status400BadRequest)]
     [Authorize(Roles = "Manager,Customer")]
     [Route("api/bookings/{id}")]
     public async Task<ActionResult<Booking>> GetBooking(int id)
     {
         var loggedId = User.FindFirstValue(ClaimTypes.SerialNumber);
         var role = User.FindFirstValue(ClaimTypes.Role);
-
+        
         try
         {
+            if (id < 1)
+            {
+                return BadRequest("Invalid ID");
+            }
+
             var booking = await _bookingService.GetBookingAsync(id, role, loggedId);
 
-            return booking != null ? Ok(booking) : StatusCode(403, "The user is not allowed to perform this action.");
+            if (booking == null)
+            {
+                return NotFound();
+            }
+
+            return Ok(booking);
         }
-        catch (Exception)
+
+        catch (InvalidOperationException ex)
         {
-            return StatusCode(StatusCodes.Status500InternalServerError, "An error occurred while getting the booking.");
+            _logger.LogError(ex, $"An error occurred while getting the booking with id = {id}");
+            return StatusCode(StatusCodes.Status500InternalServerError, ex.Message);
+        }
+        catch (ArgumentException ex)
+        {
+            return BadRequest($"An error occurred while getting the booking with id = {id}: {ex.Message}");
+        }
+        catch (Exception ex)
+        {
+            _logger.LogError(ex, $"An error occurred while getting the booking with id = {id}");
+            return StatusCode(StatusCodes.Status500InternalServerError, ex.Message);
         }
     }
 
     [HttpPut]
-    [ProducesResponseType(StatusCodes.Status202Accepted)]
-    [ProducesResponseType(StatusCodes.Status500InternalServerError)]
-    [ProducesResponseType(StatusCodes.Status403Forbidden)]
     [Authorize(Roles = "Manager,Customer")]
     [Route("api/bookings")]
-    public async Task<ActionResult> UpdateBooking([FromBody] Booking bookingEntity)
+    [ProducesResponseType(StatusCodes.Status202Accepted)]
+    [ProducesResponseType(StatusCodes.Status400BadRequest)]
+    [ProducesResponseType(StatusCodes.Status403Forbidden)]
+    [ProducesResponseType(StatusCodes.Status500InternalServerError)]
+    public async Task<ActionResult<string>> UpdateBooking([FromBody] Booking booking)
     {
+        if (booking == null)
+        {
+            return BadRequest("Booking entity cannot be null.");
+        }
+
+        if (!ModelState.IsValid)
+        {
+            return BadRequest("Invalid booking entity.");
+        }
+
         var loggedId = User.FindFirstValue(ClaimTypes.SerialNumber);
         var role = User.FindFirstValue(ClaimTypes.Role);
 
         try
         {
-            var isUpdated = await _bookingService.UpdateBookingAsync(bookingEntity, role, loggedId);
+            var isUpdated = await _bookingService.UpdateBookingAsync(booking, role, loggedId);
 
             return isUpdated
                 ? StatusCode(StatusCodes.Status202Accepted, "The booking updated.")
                 : StatusCode(StatusCodes.Status403Forbidden, "The user is not allowed to perform this action.");
         }
-        catch (Exception)
+        catch (InvalidOperationException ex)
         {
-            return StatusCode(StatusCodes.Status500InternalServerError, "An error occurred while creating the booking.");
+            _logger.LogError(ex, $"An error occurred while updating the booking with id = {booking.Id}");
+            return StatusCode(StatusCodes.Status500InternalServerError, ex.Message);
+        }
+        catch (ArgumentException ex)
+        {
+            _logger.LogError(ex, "Invalid argument.");
+            return BadRequest(ex.Message);
+        }
+        catch (Exception ex)
+        {
+            _logger.LogError(ex, $"An error occurred while updating the booking with id = {booking.Id}.");
+            return StatusCode(StatusCodes.Status500InternalServerError, ex.Message);
         }
     }
+
 
     [HttpPost]
     [ProducesResponseType(StatusCodes.Status201Created)]
@@ -116,15 +165,17 @@ public class BookingController : ControllerBase
                 ? StatusCode(StatusCodes.Status201Created, booking)
                 : StatusCode(StatusCodes.Status403Forbidden, "The user is not allowed to perform this action.");
         }
-        catch (Exception)
+        catch (Exception ex)
         {
-            return StatusCode(StatusCodes.Status500InternalServerError, "An error occurred while creating the booking.");
+            _logger.LogError(ex, $"An error occurred while creating the booking with id = {booking.Id}.");
+            return StatusCode(StatusCodes.Status500InternalServerError, ex.Message);
         }
     }
 
     [HttpDelete]
     [ProducesResponseType(StatusCodes.Status204NoContent)]
     [ProducesResponseType(StatusCodes.Status500InternalServerError)]
+    [ProducesResponseType(StatusCodes.Status400BadRequest)]
     [ProducesResponseType(StatusCodes.Status403Forbidden)]
     [Authorize(Roles = "Manager,Customer")]
     [Route("api/bookings/{id:int}")]
@@ -132,16 +183,26 @@ public class BookingController : ControllerBase
     {
         var loggedId = User.FindFirstValue(ClaimTypes.SerialNumber);
         var role = User.FindFirstValue(ClaimTypes.Role);
+    
 
         try
         {
+            if (id < 1)
+            {
+                return BadRequest("Invalid ID");
+            }
+
             bool isDeleted = await _bookingService.DeleteBookingAsync(id, role, loggedId);
 
-            return isDeleted ? NoContent() : StatusCode(403, "The user is not allowed to perform this action.");
+            return isDeleted 
+                    ? NoContent() 
+                    : StatusCode(StatusCodes.Status403Forbidden, "The user is not allowed to perform this action.");
         }
-        catch (Exception)
+        
+        catch (Exception ex)
         {
-            return StatusCode(StatusCodes.Status500InternalServerError, "An error occurred while deleting the booking.");
+            _logger.LogError(ex, $"An error occurred while deleting the booking with id = {id}");
+            return StatusCode(StatusCodes.Status500InternalServerError,ex.Message);
         }
     }
 }
